+++ NEWS.md
@@ -1,3 +1,12 @@
+# whippr 0.1.2
+
+* Fixed a bug in `read_data.cosmed()` that made the time column to return `NA` values when the test was longer than one hour.
+* Added a `custom` option to `read_data()`.
+
+# whippr 0.1.1
+
+* Updated docs with `roxygen 7.2.1`.
+
 # whippr 0.1.0
 
 * General cleanup has been performed to reduce dependencies.